+++ zappa/cli.py
@@ -34,7 +34,7 @@ import zipfile
 from click.exceptions import ClickException
 from dateutil import parser
 from datetime import datetime,timedelta
-from zappa import Zappa, logger
+from zappa import Zappa, logger, API_GATEWAY_REGIONS
 from version import __version__
 from util import check_new_version_available, detect_django_settings, detect_flask_apps
 
@@ -132,7 +132,8 @@ class ZappaCLI(object):
         Parses command, load settings and dispatches accordingly.
 
         """
-        help_message = "Please supply a command to execute. Can be one of: {}".format(', '.join(x for x in sorted(CLI_COMMANDS)))
+        help_message = ("Please supply a command to execute. " +
+                       "Can be one of: {}".format(', '.join(x for x in sorted(CLI_COMMANDS))))
 
         parser = argparse.ArgumentParser(description='Zappa - Deploy Python applications to AWS Lambda and API Gateway.\n')
         parser.add_argument('command_env', metavar='U', type=str, nargs='*', help=help_message)
@@ -144,11 +145,18 @@ class ZappaCLI(object):
                             help='The WSGI application function.')
         parser.add_argument('-v', '--version', action='store_true', help='Print the zappa version', default=False)
         parser.add_argument('-y', '--yes', action='store_true', help='Auto confirm yes', default=False)
-        parser.add_argument('--remove-logs', action='store_true', help='Removes log groups of api gateway and lambda task during the undeployment.', default=False)
-        parser.add_argument('--raw', action='store_true', help='When invoking remotely, invoke this python as a string, not as a modular path.', default=False)
-        parser.add_argument('--no-cleanup', action='store_true', help="Don't remove certificate files from /tmp during certify. Dangerous.", default=False)
-        parser.add_argument('--all', action='store_true', help="Execute this command for all of our defined Zappa environments.", default=False)
-        parser.add_argument('--json', action='store_true', help='Returns status in JSON format', default=False)  # https://github.com/Miserlou/Zappa/issues/407
+        parser.add_argument('--remove-logs', action='store_true',
+                            help='Removes log groups of api gateway and lambda task during the undeployment.',
+                            default=False)
+        parser.add_argument('--raw', action='store_true',
+                            help='When invoking remotely, invoke this python as a string, not as a modular path.',
+                            default=False)
+        parser.add_argument('--no-cleanup', action='store_true',
+                            help="Don't remove certificate files from /tmp during certify. Dangerous.", default=False)
+        parser.add_argument('--all', action='store_true',
+                            help="Execute this command for all of our defined Zappa environments.", default=False)
+        parser.add_argument('--json', action='store_true', help='Returns status in JSON format',
+                            default=False)  # https://github.com/Miserlou/Zappa/issues/407
 
         args = parser.parse_args(argv)
 
@@ -172,16 +180,16 @@ class ZappaCLI(object):
             print("The command '{}' is not recognized. {}".format(self.command, help_message))
             return
 
-        # Make sure there isn't a new version available
-        if not self.vargs['json']:
-            self.check_for_update()
-
         # We don't have any settings yet, so make those first!
         # (Settings-based interactions will fail
         # before a project has been initialized.)
         if self.command == 'init':
             self.init()
             return
+
+        # Make sure there isn't a new version available
+        if not self.vargs['json']:
+            self.check_for_update()
 
         # Load and Validate Settings File
         self.load_settings_file(self.vargs['settings_file'])
@@ -221,7 +229,8 @@ class ZappaCLI(object):
         self.api_stage = environment
 
         if command not in ['status', 'manage']:
-            click.echo("Calling " + click.style(command, fg="green", bold=True) + " for environment " + click.style(self.api_stage, bold=True) + ".." )
+            click.echo("Calling " + click.style(command, fg="green", bold=True) + " for environment " +
+                       click.style(self.api_stage, bold=True) + ".." )
 
         # Explicity define the app function.
         if self.vargs['app_function'] is not None:
@@ -269,7 +278,7 @@ class ZappaCLI(object):
             else:
                 command = command_tail[0] # ex: zappa manage dev showmigrations admin
 
-            self.invoke(command, "manage")
+            self.invoke(command, command="manage")
 
         elif command == 'tail': # pragma: no cover
             self.tail()
@@ -302,7 +311,8 @@ class ZappaCLI(object):
         # Make sure this isn't already deployed.
         deployed_versions = self.zappa.get_lambda_function_versions(self.lambda_name)
         if len(deployed_versions) > 0:
-            raise ClickException("This application is " + click.style("already deployed", fg="red") + " - did you mean to call " + click.style("update", bold=True) + "?")
+            raise ClickException("This application is " + click.style("already deployed", fg="red") +
+                                 " - did you mean to call " + click.style("update", bold=True) + "?")
 
         # Make sure the necessary IAM execution roles are available
         if self.manage_roles:
@@ -311,9 +321,11 @@ class ZappaCLI(object):
             except botocore.client.ClientError:
                 raise ClickException(
                     click.style("Failed", fg="red") + " to " + click.style("manage IAM roles", bold=True) + "!\n" +
-                    "You may " + click.style("lack the necessary AWS permissions", bold=True) + " to automatically manage a Zappa execution role.\n" +
-                    "To fix this, see here: " + click.style("https://github.com/Miserlou/Zappa#using-custom-aws-iam-roles-and-policie", bold=True) + '\n'
-                )
+                    "You may " + click.style("lack the necessary AWS permissions", bold=True) +
+                    " to automatically manage a Zappa execution role.\n" +
+                    "To fix this, see here: " +
+                    click.style("https://github.com/Miserlou/Zappa#using-custom-aws-iam-roles-and-policie", bold=True)
+                    + '\n')
 
         # Create the Lambda Zip
         self.create_package()
@@ -398,11 +410,14 @@ class ZappaCLI(object):
             last_updated = parser.parse(conf['LastModified'])
             last_updated_unix = time.mktime(last_updated.timetuple())
         except Exception as e:
-            click.echo(click.style("Warning!", fg="red") + " Couldn't get function - have you deployed yet?")
+            click.echo(click.style("Warning!", fg="red") + " Couldn't get function " + self.lambda_name +
+                       " in " + self.zappa.aws_region + " - have you deployed yet?")
             sys.exit(-1)
 
         if last_updated_unix <= updated_time:
-            click.echo(click.style("Warning!", fg="red") + " You may have upgraded Zappa since deploying this application. You will need to " + click.style("redeploy", bold=True) + " for this deployment to work properly!")
+            click.echo(click.style("Warning!", fg="red") +
+                       " You may have upgraded Zappa since deploying this application. You will need to " +
+                       click.style("redeploy", bold=True) + " for this deployment to work properly!")
 
         # Make sure the necessary IAM execution roles are available
         if self.manage_roles:
@@ -410,8 +425,11 @@ class ZappaCLI(object):
                 self.zappa.create_iam_roles()
             except botocore.client.ClientError:
                 click.echo(click.style("Failed", fg="red") + " to " + click.style("manage IAM roles", bold=True) + "!")
-                click.echo("You may " + click.style("lack the necessary AWS permissions", bold=True) + " to automatically manage a Zappa execution role.")
-                click.echo("To fix this, see here: " + click.style("https://github.com/Miserlou/Zappa#using-custom-aws-iam-roles-and-policie", bold=True))
+                click.echo("You may " + click.style("lack the necessary AWS permissions", bold=True) +
+                           " to automatically manage a Zappa execution role.")
+                click.echo("To fix this, see here: " +
+                           click.style("https://github.com/Miserlou/Zappa#using-custom-aws-iam-roles-and-policies",
+                                       bold=True))
                 sys.exit(-1)
 
         # Create the Lambda Zip,
@@ -603,7 +621,8 @@ class ZappaCLI(object):
             timeout = conf['Timeout']
 
             if timeout < 60:
-                click.echo(click.style("Unable to schedule certificate autorenewer!", fg="red", bold=True) + " Please redeploy with a " + click.style("timeout_seconds", bold=True) + " greater than 60!")
+                click.echo(click.style("Unable to schedule certificate autorenewer!", fg="red", bold=True) +
+                           " Please redeploy with a " + click.style("timeout_seconds", bold=True) + " greater than 60!")
             else:
                 events.append({'name': 'zappa-le-certify',
                                'function': 'handler.certify_callback',
@@ -614,7 +633,9 @@ class ZappaCLI(object):
             try:
                 function_response = self.zappa.lambda_client.get_function(FunctionName=self.lambda_name)
             except botocore.exceptions.ClientError as e: # pragma: no cover
-                click.echo(click.style("Function does not exist", fg="yellow") + ", please " + click.style("deploy", bold=True) + "first. Ex:" + click.style("zappa deploy {}.".format(self.api_stage), bold=True))
+                click.echo(click.style("Function does not exist", fg="yellow") + ", please " +
+                           click.style("deploy", bold=True) + "first. Ex:" +
+                           click.style("zappa deploy {}.".format(self.api_stage), bold=True))
                 sys.exit(-1)
 
             print("Scheduling..")
@@ -655,16 +676,20 @@ class ZappaCLI(object):
             )
 
 
-    def invoke(self, function_name, raw_python=False):
+    def invoke(self, function_name, raw_python=False, command=None):
         """
         Invoke a remote function.
         """
 
-        # Invoke it!
+        # There are three likely scenarios for 'command' here:
+        #   command, which is a modular function path
+        #   raw_command, which is a string of python to execute directly
+        #   manage, which is a Django-specific management command invocation
+        key = command if command is not None else 'command'
         if raw_python:
             command = {'raw_command': function_name}
         else:
-            command = {'command': function_name} # This is really a string
+            command = {key: function_name}
 
         # Can't use hjson
         import json as json
@@ -696,7 +721,8 @@ class ZappaCLI(object):
         lambda_versions = self.zappa.get_lambda_function_versions(self.lambda_name)
 
         if not lambda_versions:
-            raise ClickException(click.style("No Lambda detected - have you deployed yet?", fg='red'))
+            raise ClickException(click.style("No Lambda %s detected in %s - have you deployed yet?" %
+                                             (self.lambda_name, self.zappa.aws_region), fg='red'))
 
         status_dict = collections.OrderedDict()
         status_dict["Lambda Versions"] = len(lambda_versions)
@@ -864,13 +890,16 @@ class ZappaCLI(object):
 ╚══════╝╚═╝  ╚═╝╚═╝     ╚═╝     ╚═╝  ╚═╝\n""", fg='green', bold=True))
 
         click.echo(click.style("Welcome to ", bold=True) + click.style("Zappa", fg='green', bold=True) + click.style("!\n", bold=True))
-        click.echo(click.style("Zappa", bold=True) + " is a system for running server-less Python web applications on AWS Lambda and AWS API Gateway.")
+        click.echo(click.style("Zappa", bold=True) + " is a system for running server-less Python web applications"
+                                                     " on AWS Lambda and AWS API Gateway.")
         click.echo("This `init` command will help you create and configure your new Zappa deployment.")
         click.echo("Let's get started!\n")
 
         # Create Env
         while True:
-            click.echo("Your Zappa configuration can support multiple production environments, like '" + click.style("dev", bold=True)  + "', '" + click.style("staging", bold=True)  + "', and '" + click.style("production", bold=True)  + "'.")
+            click.echo("Your Zappa configuration can support multiple production environments, like '" +
+                       click.style("dev", bold=True)  + "', '" + click.style("staging", bold=True)  + "', and '" +
+                       click.style("production", bold=True)  + "'.")
             env = raw_input("What do you want to call this environment (default 'dev'): ") or "dev"
             try:
                 self.check_stage_name(env)
@@ -889,13 +918,13 @@ class ZappaCLI(object):
         try: # pragma: no cover
             import django
             has_django = True
-        except ImportError, e:
+        except ImportError as e:
             has_django = False
 
         try: # pragma: no cover
             import flask
             has_flask = True
-        except ImportError, e:
+        except ImportError as e:
             has_flask = False
 
         print('')
@@ -934,17 +963,59 @@ class ZappaCLI(object):
 
         # TODO: Create VPC?
         # Memory size? Time limit?
+        # Domain? LE keys? Region?
+        # 'Advanced Settings' mode?
 
-        # Confirm
-        zappa_settings = {
-            env: {
-                's3_bucket': bucket,
-            }
-        }
-        if has_django:
-            zappa_settings[env]['django_settings'] = django_settings
+        # Globalize
+        click.echo("\nYou can optionally deploy to " + click.style("all available regions", bold=True)  + " in order to provide fast global service.")
+        click.echo("If you are using Zappa for the first time, you probably don't want to do this!")
+        global_deployment = False
+        while True:
+            global_type = raw_input("Would you like to deploy this application to " + click.style("globally", bold=True)  + "? (default 'n') [y/n/(p)rimary]: ")
+            if not global_type:
+                break
+            if global_type.lower() in ["y", "yes", "p", "primary"]:
+                global_deployment = True
+                break
+            if global_type.lower() in ["n", "no"]:
+                global_deployment = False
+                break
+
+        if global_deployment:
+            regions = API_GATEWAY_REGIONS
+            if global_type.lower() in ["p", "primary"]:
+                envs = [{env + '_' + region.replace('-', '_'): { 'aws_region': region}} for region in regions if '-1' in region]
+            else:
+                envs = [{env + '_' + region.replace('-', '_'): { 'aws_region': region}} for region in regions]
         else:
-            zappa_settings[env]['app_function'] = app_function
+            region = None # assume system default
+            envs = [{env: {}}]
+
+        zappa_settings = {}
+        for each_env in envs:
+
+            # Honestly, this could be cleaner.
+            env_name = each_env.keys()[0]
+            env_dict = each_env[env_name]
+
+            env_bucket = bucket
+            if global_deployment:
+                env_bucket = bucket.replace('-', '_') + '_' + env_name
+
+            env_zappa_settings = {
+                env_name: {
+                    's3_bucket': env_bucket,
+                }
+            }
+            if env_dict.has_key('aws_region'):
+                env_zappa_settings[env_name]['aws_region'] = env_dict.get('aws_region')
+
+            zappa_settings.update(env_zappa_settings)
+
+            if has_django:
+                zappa_settings[env_name]['django_settings'] = django_settings
+            else:
+                zappa_settings[env_name]['app_function'] = app_function
 
         import json as json # hjson is fine for loading, not fine for writing.
         zappa_settings_json = json.dumps(zappa_settings, sort_keys=True, indent=4)
@@ -952,7 +1023,7 @@ class ZappaCLI(object):
         click.echo("\nOkay, here's your " + click.style("zappa_settings.js", bold=True) + ":\n")
         click.echo(click.style(zappa_settings_json, fg="yellow", bold=False))
 
-        confirm = raw_input("\nDoes this look " + click.style("okay", bold=True, fg="green")  + "? (default y) [y/n]: ") or 'yes'
+        confirm = raw_input("\nDoes this look " + click.style("okay", bold=True, fg="green")  + "? (default 'y') [y/n]: ") or 'yes'
         if confirm[0] not in ['y', 'Y', 'yes', 'YES']:
             click.echo("" + click.style("Sorry", bold=True, fg='red') + " to hear that! Please init again.")
             return
@@ -961,14 +1032,23 @@ class ZappaCLI(object):
         with open("zappa_settings.json", "w") as zappa_settings_file:
             zappa_settings_file.write(zappa_settings_json)
 
-        click.echo("\n" + click.style("Done", bold=True) + "! Now you can " + click.style("deploy", bold=True)  + " your Zappa application by executing:\n")
-        click.echo(click.style("\t$ zappa deploy %s" % env, bold=True))
+        if global_deployment:
+            click.echo("\n" + click.style("Done", bold=True) + "! You can also " + click.style("deploy all", bold=True)  + " by executing:\n")
+            click.echo(click.style("\t$ zappa deploy --all", bold=True))
 
-        click.echo("\nAfter that, you can " + click.style("update", bold=True) + " your application code with:\n")
-        click.echo(click.style("\t$ zappa update %s" % env, bold=True))
+            click.echo("\nAfter that, you can " + click.style("update", bold=True) + " your application code with:\n")
+            click.echo(click.style("\t$ zappa update --all", bold=True))
+        else:
+            click.echo("\n" + click.style("Done", bold=True) + "! Now you can " + click.style("deploy", bold=True)  + " your Zappa application by executing:\n")
+            click.echo(click.style("\t$ zappa deploy %s" % env, bold=True))
 
-        click.echo("\nTo learn more, check out our project page on " + click.style("GitHub", bold=True) + " here: " + click.style("https://github.com/Miserlou/Zappa", fg="cyan", bold=True))
-        click.echo("and stop by our " + click.style("Slack", bold=True) + " channel here: " + click.style("http://bit.do/zappa", fg="cyan", bold=True))
+            click.echo("\nAfter that, you can " + click.style("update", bold=True) + " your application code with:\n")
+            click.echo(click.style("\t$ zappa update %s" % env, bold=True))
+
+        click.echo("\nTo learn more, check out our project page on " + click.style("GitHub", bold=True) +
+                   " here: " + click.style("https://github.com/Miserlou/Zappa", fg="cyan", bold=True))
+        click.echo("and stop by our " + click.style("Slack", bold=True) + " channel here: " +
+                   click.style("http://bit.do/zappa", fg="cyan", bold=True))
         click.echo("\nEnjoy!,")
         click.echo(" ~ Team " + click.style("Zappa", bold=True) + "!")
 
@@ -982,14 +1062,17 @@ class ZappaCLI(object):
         # Give warning on --no-cleanup
         if no_cleanup:
             clean_up = False
-            click.echo(click.style("Warning!", fg="yellow", bold=True) + " You are calling certify with " + click.style("--no-cleanup", bold=True) + ". Your certificate files will remain in the system temporary directory after this command exectutes!")
+            click.echo(click.style("Warning!", fg="yellow", bold=True) + " You are calling certify with " +
+                       click.style("--no-cleanup", bold=True) +
+                       ". Your certificate files will remain in the system temporary directory after this command exectutes!")
         else:
             clean_up = True
 
         # Make sure this isn't already deployed.
         deployed_versions = self.zappa.get_lambda_function_versions(self.lambda_name)
         if len(deployed_versions) == 0:
-            raise ClickException("This application " + click.style("isn't deployed yet", fg="red") + " - did you mean to call " + click.style("deploy", bold=True) + "?")
+            raise ClickException("This application " + click.style("isn't deployed yet", fg="red") +
+                                 " - did you mean to call " + click.style("deploy", bold=True) + "?")
 
         # Get install account_key to /tmp/account_key.pem
         account_key_location = self.stage_config.get('lets_encrypt_key')
@@ -1015,9 +1098,9 @@ class ZappaCLI(object):
                 from shutil import copyfile
                 copyfile(account_key_location, '/tmp/account.key')
         else:
-
             if not cert_location or not cert_key_location or not cert_chain_location:
-                raise ClickException("Can't certify a domain without " + click.style("certificate, certificate_key and certificate_chain", fg="red", bold=True) + " configured!")
+                raise ClickException("Can't certify a domain without " +
+                                     click.style("certificate, certificate_key and certificate_chain", fg="red", bold=True) + " configured!")
 
             # Read the supplied certificates.
             with open(cert_location) as f:
@@ -1053,7 +1136,8 @@ class ZappaCLI(object):
                     self.lambda_name,
                     self.api_stage
                 )
-                print("Created a new domain name. Please note that it can take up to 40 minutes for this domain to be created and propagated through AWS, but it requires no further work on your part.")
+                print("Created a new domain name. Please note that it can take up to 40 minutes for this domain to be "
+                      "created and propagated through AWS, but it requires no further work on your part.")
             else:
                 self.zappa.update_domain_name(
                     domain,
@@ -1105,9 +1189,11 @@ class ZappaCLI(object):
             version = __version__
             updateable = check_new_version_available(version)
             if updateable:
-                click.echo(click.style("Important!", fg="yellow", bold=True) + " A new version of " + click.style("Zappa", bold=True) + " is available!")
+                click.echo(click.style("Important!", fg="yellow", bold=True) +
+                           " A new version of " + click.style("Zappa", bold=True) + " is available!")
                 click.echo("Upgrade with: " + click.style("pip install zappa --upgrade", bold=True))
-                click.echo("Visit the project page on GitHub to see the latest changes: " + click.style("https://github.com/Miserlou/Zappa", bold=True))
+                click.echo("Visit the project page on GitHub to see the latest changes: " +
+                           click.style("https://github.com/Miserlou/Zappa", bold=True))
         except Exception as e: # pragma: no cover
             print(e)
             return
@@ -1146,7 +1232,8 @@ class ZappaCLI(object):
             self.project_name = slugify.slugify(os.getcwd().split(os.sep)[-1])[:15]
 
         if len(self.project_name) > 15: # pragma: no cover
-            click.echo(click.style("Warning", fg="red", bold=True) + "! Your " + click.style("project_name", bold=True) + " may be too long to deploy! Please make it <16 characters.")
+            click.echo(click.style("Warning", fg="red", bold=True) + "! Your " + click.style("project_name", bold=True) +
+                       " may be too long to deploy! Please make it <16 characters.")
 
         # The name of the actual AWS Lambda function, ex, 'helloworld-dev'
         # Django's slugify doesn't replace _, but this does.
@@ -1164,7 +1251,7 @@ class ZappaCLI(object):
         self.exception_handler = self.zappa_settings[
             self.api_stage].get('exception_handler', None)
         self.aws_region = self.zappa_settings[
-            self.api_stage].get('aws_region', 'us-east-1')
+            self.api_stage].get('aws_region', None)
         self.debug = self.zappa_settings[
             self.api_stage].get('debug', True)
         self.prebuild_script = self.zappa_settings[
@@ -1222,7 +1309,9 @@ class ZappaCLI(object):
         if self.app_function:
             self.collision_warning(self.app_function)
             if self.app_function[-3:] == '.py':
-                click.echo(click.style("Warning!", fg="yellow", bold=True) + " Your app_function is pointing to a " + click.style("file and not a function", bold=True) + "! It should probably be something like 'my_file.app', not 'my_file.py'!")
+                click.echo(click.style("Warning!", fg="yellow", bold=True) +
+                           " Your app_function is pointing to a " + click.style("file and not a function", bold=True) +
+                           "! It should probably be something like 'my_file.app', not 'my_file.py'!")
 
         return self.zappa
 
@@ -1297,8 +1386,12 @@ class ZappaCLI(object):
                 )
 
             # Local envs
+            env_dict = {}
+            if self.aws_region:
+                env_dict['AWS_REGION'] = self.aws_region
+            env_dict.update(dict(self.environment_variables))
             settings_s = settings_s + "ENVIRONMENT_VARIABLES={0}\n".format(
-                    dict(self.environment_variables)
+                    env_dict
                 )
 
             # We can be environment-aware
@@ -1417,7 +1510,9 @@ class ZappaCLI(object):
         ]
         for namespace_collision in namespace_collisions:
             if namespace_collision in item:
-                click.echo(click.style("Warning!", fg="yellow", bold=True) + " You may have a namespace collision with " + click.style(item, bold=True) + "! You may want to rename that file.")
+                click.echo(click.style("Warning!", fg="yellow", bold=True) +
+                           " You may have a namespace collision with " + click.style(item, bold=True) +
+                           "! You may want to rename that file.")
 
     def deploy_api_gateway(self, api_id):
         cache_cluster_enabled = self.stage_config.get('cache_cluster_enabled', False)
@@ -1443,9 +1538,13 @@ def shamelessly_promote():
     Shamelessly promote our little community.
     """
 
-    click.echo("Need " + click.style("help", fg='green', bold=True) + "? Found a " + click.style("bug", fg='green', bold=True) + "? Let us " + click.style("know", fg='green', bold=True) + "! :D")
-    click.echo("File bug reports on " + click.style("GitHub", bold=True) + " here: " + click.style("https://github.com/Miserlou/Zappa", fg='cyan', bold=True))
-    click.echo("And join our " + click.style("Slack", bold=True) + " channel here: " + click.style("https://slack.zappa.io", fg='cyan', bold=True))
+    click.echo("Need " + click.style("help", fg='green', bold=True) +
+               "? Found a " + click.style("bug", fg='green', bold=True) +
+               "? Let us " + click.style("know", fg='green', bold=True) + "! :D")
+    click.echo("File bug reports on " + click.style("GitHub", bold=True) + " here: "
+               + click.style("https://github.com/Miserlou/Zappa", fg='cyan', bold=True))
+    click.echo("And join our " + click.style("Slack", bold=True) + " channel here: "
+               + click.style("https://slack.zappa.io", fg='cyan', bold=True))
     click.echo("Love!,")
     click.echo(" ~ Team " + click.style("Zappa", bold=True) + "!")